fix(security): expand PHPStan coverage, fix ModSecurity grep, boost test coverage to 90% (#104)
Security audit findings:
- Fix binary file detection in ModSecurity grep (add -a flag)
- Fix XSS in email template (escape user params with e())
- Fix null-safety in all middlewares (Auth::user() null checks)
- Fix type errors in Jobs, Services, Models, Listeners

PHPStan:
- Expand from 3 paths to full app/ directory
- Fix all property.nonObject and method.nonObject errors (security-critical)
- Add PHPDoc types to all Models, Console Commands
- Generate baseline for 110 remaining missingType.* errors
- Exclude interactive CLI commands (Laravel\Prompts) and Filament/Livewire

Test coverage 78.6% → 90.2%:
- Add 79 new tests (Exceptions, Middlewares, Models, Services, Traits)
- Add reasonable exclusions for untestable code (SSH, external APIs, interactive CLI)

Other:
- Install PHPInsights (nunomaduro/phpinsights)
- Create .security-audit.json config
- Fix config comment (10 days → 7 days for report_expiration)
- Apply Pint formatting

+++ app/Services/PatternDetection/AnomalyDetector.php
@@ -121,7 +121,7 @@ class AnomalyDetector
      */
     private function getCurrentTraffic(): int
     {
-        return DB::table('ip_reputation')
+        return (int) DB::table('ip_reputation')
             ->where('last_seen_at', '>=', now()->subHour())
             ->sum('total_requests');
     }
